app/requirements.txt
streamlit==1.43.2
openai==1.66.3
PyPDF2==3.0.1
python-dotenv==1.0.0
tqdm==4.67.1
nest_asyncio==1.6.0
requests==2.32.3
